requirements.txt
requests
beautifulsoup4
markdownify
openai
python-dotenv
genanki
InquirerPy

